优化视频号订单相关功能；测试视频号电子面单打印

+++ open-api/wei-api/src/main/java/com/qihang/wei/service/SShopService.java
@@ -9,7 +9,7 @@ import com.baomidou.mybatisplus.extension.service.IService;
 * @createDate 2024-06-03 14:14:56
 */
 public interface SShopService extends IService<SShop> {
-    SShop selectShopById(Integer shopId);
+    SShop selectShopById(Long shopId);
 
-    void updateSessionKey(Integer shopId,String sessionKey);
+    void updateSessionKey(Long shopId,String sessionKey);
 }
